Move asyncCall breakpoint handler into a separate file
